+++ src/main/utils/time_utils.py
@@ -1,0 +1,5 @@
+import time
+
+
+def current_milli_time():
+    return round(time.time() * 1000)
